Creating my landing page:
Started with creating the layout for the html page. Now slowly adding
the css to it to make it look better. Starting from the top of the page
and making my way down. So far, I've completed the header part of css. I
would like to complete the links and make them interactive.