Changed my portfolio page
I changed the profile pic and used a photo that I stripped the
background from. I also added fonts as well to stylize the page a bit
more. Something that I learned is that when I first created this page
the divs and layouts aren't the cleanest. I tried rearranging the css
and html code to make it easier to read. Will work on this more later.

Now I want to put an about me section and push that above the "Cool
Stuff I've created" section.