Configuring the a new portfolio/landing page
Grabbed a template I found on bootstrap to see how it works so far. So
far I like the template and how easy it is to manipulate the code. I
have changed the header, profile picture, and one of the portfolio
pieces for one of my project called BlueBicho.

I plan on working on this more and if I like it, merge the branch to the
main branch.