A lot of updates
I changed my profile picture to a chibi version of me. I also added a
logo as well, but I'm thinking about changing it. I removed some items
from the portfolio section, only leaving two items. In the footer, I
removed a section that was referencing freelancing that came with the
original template. I also made the links in the Around the Wrb field
point to my social media accounts. I've also removed other links that
were there before. I haven't put my resume up there yet because I'm
still working on it. Most importantly, I've added a snippet of code from
Calendly that allows people to schedule a meeting with me. It is on the
bottom of the page and it is stays there as you scroll down.

To stay on track, I just need to do the following:
1. Upload my resume.
2. Change the logo.

Long term goals:
1. Create more projects to add to the portfolio.
2. Add a "how many people visit the site" counter and display it.
3. Make the form on the site work so people can email me through it.