Minor Tweaks
I changed the + symbol when someone hovers over an portfolio item to the
name of what the portfolio item is. I also added a resume to the website
that is in a pdf format. Thank you, Chris Bogner for the idea.

Also, when changed the stars to my logo within the individual portfolio
items.