Minor tweaks
I added a few minor tweaks to the index.html file. I converted my resume
from a .docx file to a .pdf file and added to the download resume
button. I also removed the + sign from the portfolio section when you
hover over it. Instead, I replaced them with the name of the projects.
Thank you Chris Bogner for the suggestion.

I also changed the star symbols with the portfolio section to have the
MBC (Matt Blackburn Codes) logo instead.