Service Offerings addition
Feat: Added a section for new services offerings.
The section uses a table to display the services,
cost, and a brief description of each service.

Feat: Added service offerings to the menu bar. By
clicking on the service offerings link, the user
will be taken to the service offerings section.