Updated portfolio images and updated information.
Removed Blue Bicho completely and just added the
individual projects to the portfolio. For example,
Papcorn, Mako2, and Blastroids. Also updated the
images for the projects and archived the old
images.

Also, added more text to create a more inviting
experience for the user. underneath the about and
contact sections.